Renamed CubeData to LogData
